personal-ai-headhunter/requirements.txt
streamlit
pandas
sqlalchemy
psycopg2-binary
openai
chromadb>=1.4.0
python-dotenv
openpyxl
tiktoken
plotly


